Add files via upload
The Titanic dataset is a classic dataset used in data science and machine learning to explore binary classification tasks. It contains real historical data about the passengers aboard the RMS Titanic, which sank after hitting an iceberg on April 15, 1912.

The goal is to predict whether a passenger survived or not based on features like age, gender, class, fare, and more.